basket domain tests are adding ...

+++ src/Services/Basket/Basket.Application/UseCases/DeleteBasket/DeleteBasketCommandHandler.cs
@@ -1,5 +1,7 @@
+using System.Runtime.InteropServices.WindowsRuntime;
 using System.Threading;
 using System.Threading.Tasks;
+using Basket.Domain.Commands;
 using Basket.Domain.Commands.DeleteBasket;
 using Basket.Domain.RepositoryInterfaces;
 using MediatR;
@@ -18,8 +20,29 @@ namespace Basket.Application.UseCases.DeleteBasket
         public async Task<DeleteBasketCommandResult> Handle(DeleteBasketCommand request,
             CancellationToken cancellationToken)
         {
-            await _basketCommandRepository.DeleteAsync(request.BuyerId, cancellationToken);
-            return new DeleteBasketCommandResult();
+            if (request == null)
+            {
+                return new DeleteBasketCommandResult()
+                {
+                    ValidateState = ValidationState.NotAcceptable,
+                    ReturnPath = "/basket"
+                };
+            }
+
+            var isDeleted = await _basketCommandRepository.DeleteAsync(request.BuyerId, cancellationToken);
+            if (!isDeleted)
+            {
+                return new DeleteBasketCommandResult()
+                {
+                    ValidateState = ValidationState.DoesNotExist,
+                    ReturnPath = "/basket"
+                };
+            }
+
+            return new DeleteBasketCommandResult()
+            {
+                ValidateState = ValidationState.Valid
+            };
         }
     }
 }
